requirements.txt
pytorch-gpu
pillow
torchvision
numpy
matplotlib
faiss-gpu
tqdm
